feat: 관리자 페이지에 공유하기 버튼 추가
