+++ client/src/features/Event/Manage/components/info/EventHeader.tsx
@@ -1,7 +1,5 @@
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
-import { HttpError } from '@/api/fetcher';
-import { useCloseEventRegistration } from '@/api/mutations/useCloseEventRegistration';
 import { badgeText } from '@/features/Event/utils/badgeText';
 import { Badge } from '@/shared/components/Badge';
 import { Button } from '@/shared/components/Button';
@@ -9,10 +7,7 @@ import { Flex } from '@/shared/components/Flex';
 import { Icon } from '@/shared/components/Icon';
 import { Text } from '@/shared/components/Text';
 import { useToast } from '@/shared/components/Toast/ToastContext';
-import { useModal } from '@/shared/hooks/useModal';
 import { formatDate } from '@/shared/utils/dateUtils';
-
-import { DeadlineModal } from './DeadlineModal';
 
 type EventHeaderProps = {
   eventId: number;
@@ -31,28 +26,13 @@ export const EventHeader = ({
   eventEnd,
   registrationEnd,
 }: EventHeaderProps) => {
-  const { success, error } = useToast();
-  const navigate = useNavigate();
+  const { success } = useToast();
   const { organizationId } = useParams();
-  const isClosed = registrationEnd ? new Date(registrationEnd) < new Date() : false;
-  const { isOpen, open, close } = useModal();
-
   const badgeTextValue = badgeText(registrationEnd);
 
-  const { mutate: closeEventRegistrationMutate } = useCloseEventRegistration(eventId);
-
-  const handleDeadlineChangeClick = () => {
-    closeEventRegistrationMutate(undefined, {
-      onSuccess: () => {
-        success('이벤트가 마감되었습니다.');
-        close();
-      },
-      onError: (err) => {
-        if (err instanceof HttpError) {
-          error(err.message);
-        }
-      },
-    });
+  const handleShareClick = () => {
+    navigator.clipboard.writeText(`${window.location.origin}/${organizationId}/event/${eventId}`);
+    success('공유 링크가 복사되었습니다.');
   };
 
   return (
@@ -60,25 +40,9 @@ export const EventHeader = ({
       <Flex dir="column" gap="12px">
         <Flex dir="row" justifyContent="space-between" alignItems="flex-end">
           <Badge variant={badgeTextValue.color}>{badgeTextValue.text}</Badge>
-          {isClosed ? (
-            <Button size="sm" color="tertiary" variant="solid" disabled>
-              마감됨
-            </Button>
-          ) : (
-            <Flex alignItems="center" gap="8px">
-              <Button
-                size="md"
-                color="primary"
-                variant="solid"
-                onClick={() => navigate(`/${organizationId}/event/${eventId}/edit`)}
-              >
-                수정하기
-              </Button>
-              <Button size="md" color="secondary" variant="solid" onClick={open}>
-                마감하기
-              </Button>
-            </Flex>
-          )}
+          <Button size="md" color="secondary" variant="solid" onClick={handleShareClick}>
+            공유하기
+          </Button>
         </Flex>
         <Text as="h1" type="Display" weight="bold">
           {title}
@@ -104,7 +68,6 @@ export const EventHeader = ({
           </Flex>
         </Flex>
       </Flex>
-      <DeadlineModal isOpen={isOpen} onClose={close} onDeadlineChange={handleDeadlineChangeClick} />
     </>
   );
 };
